vci: rewrite defaults file only if changed
The defaults file /etc/default/kdump-tools get overwritten for every
configuration change that enables kernel crash dump. This file also
gets rewritten on each reboot.

In some cases these rewrites wasn't flushed to disk before a kernel
crash. The crash dump is captured correctly, but on the boot after
crash dump, the vyatta-kdump-load service fails to load because of
the empty defaults file

The changes in this commit ensures that defaults file is written only if
the file contents have changed. After writing, the file is explicitly
flushed to the disk.

VRVDR-55751